fixing style cop issues
