+++ src/Messenger.Simulator/SimReceiver.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="SimReciever.cs" company="Mark Lauter">
+﻿// <copyright file="SimReceiver.cs" company="Mark Lauter">
 // Copyright (c) Mark Lauter. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 // </copyright>
