+++ src/Tello.App.UWP/Properties/AssemblyInfo.cs
@@ -4,7 +4,6 @@
 // </copyright>
 
 using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following
