+++ src/Tello.App/MvvM/IInputCommand.cs
@@ -1,0 +1,14 @@
+﻿// <copyright file="IInputCommand.cs" company="Mark Lauter">
+// Copyright (c) Mark Lauter. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+// </copyright>
+
+using System.Windows.Input;
+
+namespace Tello.App.MvvM
+{
+    public interface IInputCommand : ICommand
+    {
+        bool Enabled { get; set; }
+    }
+}
