+++ src/Tello.App/MvvM/IInputCommand{T}.cs
@@ -1,0 +1,13 @@
+﻿// <copyright file="IInputCommand{T}.cs" company="Mark Lauter">
+// Copyright (c) Mark Lauter. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+// </copyright>
+
+using System.Windows.Input;
+
+namespace Tello.App.MvvM
+{
+    public interface IInputCommand<T> : IInputCommand
+    {
+    }
+}
