+++ src/Tello.App/MvvM/ViewModel.cs
@@ -4,32 +4,9 @@
 // </copyright>
 
 using System;
-using System.Collections.Generic;
 
 namespace Tello.App.MvvM
 {
-    public class OpenEventArgs : EventArgs
-    {
-        public OpenEventArgs()
-            : this(null)
-        {
-        }
-
-        public OpenEventArgs(Dictionary<string, object> args)
-        {
-            this.Args = args != null
-                ? new Dictionary<string, object>(args)
-                : new Dictionary<string, object>();
-        }
-
-        public Dictionary<string, object> Args { get; }
-    }
-
-    public class ClosingEventArgs : EventArgs
-    {
-        public bool CanClose { get; set; }
-    }
-
     public class ViewModel : PropertyChangedNotifier
     {
         protected readonly IUINotifier UserNotifier;
