+++ src/Video.Sandbox/Video.cs
@@ -3,28 +3,24 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 // </copyright>
 
-using System;
 using System.Runtime.InteropServices;
-using Windows.Media.Capture.Frames;
 
 namespace Video.Sandbox
 {
     public class Video
     {
         [DllImport("/cpp/welsdec", CallingConvention = CallingConvention.Cdecl, SetLastError = true, EntryPoint = "WelsCreateDecoder")]
-        static extern void WelsCreateDecoder([MarshalAs(UnmanagedType.LPStr)]string data);
+        private static extern void WelsCreateDecoder([MarshalAs(UnmanagedType.LPStr)]string data);
 
         [DllImport("/cpp/welsdec", CallingConvention = CallingConvention.Cdecl, SetLastError = true, EntryPoint = "WelsDestroyDecoder")]
-        static extern void WelsDestroyDecoder([MarshalAs(UnmanagedType.LPStr)]string data);
+        private static extern void WelsDestroyDecoder([MarshalAs(UnmanagedType.LPStr)]string data);
 
         [DllImport("/cpp/welsdec", CallingConvention = CallingConvention.Cdecl, SetLastError = true, EntryPoint = "WelsGetDecoderCapability")]
-        static extern void WelsGetDecoderCapability([MarshalAs(UnmanagedType.LPStr)]string data);
+        private static extern void WelsGetDecoderCapability([MarshalAs(UnmanagedType.LPStr)]string data);
 
-
-        public void f()
+        public void Foo()
         {
             // var mfr = new MediaFrameReader();
-
         }
     }
 }
